requirements.txt
fastapi
uvicorn
python-dotenv
requests
rich
pydantic
sqlite-utils
scikit-learn
numpy